+++ src/accelerate/test_utils/testing.py
@@ -25,8 +25,9 @@ from unittest import mock
 
 import torch
 
+from ..operations import gather
 from ..state import AcceleratorState, is_tpu_available
-from ..utils import gather, is_tensorflow_available
+from ..utils import is_tensorflow_available
 
 
 def parse_flag_from_env(key, default=False):
